fix: Ensure that tesselating an unprojectable edge does not result in a crash (#47)
Closes #43.

This PR updates the orthographic projection to return nan, nan when
attempting to project a point that would otherwise not be visible on
that hemisphere of the earth. Allowing such a point to be projected led
to an unstable situtation when points were repeatedly roundtripped
(e.g., in edge tessellation).

This also applies formatting and fixes a compiler warning (I haven't yet
set up a pre-commit!).