feat: Implement more functions as Arrow scalar UDFs (#64)
This PR implements ArrowUDFs on top of the existing
`Geography`/`GeographyIndex` related functions. I didn't make an attempt
to optimize anything for the vectorized WKB input case yet, although
there's significant opportunity to do so (e.g., no need for a
`Geography` to calculate length or perimeter).

The side effect of this is that I finally wrote tests that live in this
repo! Eventually this will cover a range of postgis/bigquery-alike
functions, although these are the main ones. The nice part about using
WKB as input/output is that a number of other functions (e.g., WKT/WKB
IO) don't need specialized implementations for geography. The bad part
is obviously that you pay the cost of IO every time you want to call a
function. There's a few ways to mitigate this (including adding a
specialized Arrow input type that's a wrapper around the s2-specific
serialization.